Multiple changes to address issues #2 , and #3
Multiple changes to address issues #2 , and #3
removed download card call, and instead just call urllib.request directly. This way, a variable can be passed to it for the card quanitity a bit easier.

Created a "get_all_card" function that serves as a better way to get all the card data from a deck Allows for capturing the card name, and improves speed of the program.